feat: move ISO 2533/ADD 1:1985 code into its own class
